pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pave-agent"
version = "0.1.0"
description = "Semiconductor PDK Cell-level PPA analysis chatbot agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0",
    "litellm>=1.0.0",
    "oracledb>=2.0.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "matplotlib>=3.8.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
